Add MarriageAbroadCalculator#country_name_uppercase_prefix
I've moved the behaviour from the `country_name_uppercase_prefix`
calculate block to this new method.

It seems slightly odd to me that the implementation of this method is
different to `#country_name_lowercase_prefix` but that'll have to wait
for another time.

I've used stub.quacks_like(<instance-of-object>) in some of these tests.
This is instead of using a real instance of the object, or of using a
plain stub. It gives me confidence that I'm not accidentally relying on
behaviour in the real object, and that I'm not stubbing/calling a method
that doesn't exist on the real object.