Add MarriageAbroadCalculator#country_name_partner_residence
I've moved the behaviour from the `country_name_partner_residence`
calculate block to this new method.

I've used stub.quacks_like(<instance-of-object>) in some of these tests.
This is instead of using a real instance of the object, or of using a
plain stub. It gives me confidence that I'm not accidentally relying on
behaviour in the real object, and that I'm not stubbing/calling a method
that doesn't exist on the real object.