Add MarriageAbroadCalculator#embassy_or_consulate_ceremony_country
I've moved the behaviour from the
`embassy_or_consulate_ceremony_country` calculate block to this new
method.

I've used stub.quacks_like(<instance-of-object>) in some of these tests.
This is instead of using a real instance of the object, or of using a
plain stub. It gives me confidence that I'm not accidentally relying on
behaviour in the real object, and that I'm not stubbing/calling a method
that doesn't exist on the real object.