Update question for student finance forms
With full-time students no longer eligible to
apply for finance for the 14-15 year, this commit
splits the what_year question into one for
part time students and one for full time students
allowing independent values to be shown to the
user to select from.

All full-time related outcomes are updated to
reflect the new year option, removing year-1415
and replacing with year-1617.