Add #opposite_sex_no_consular_cni_country?
I want the `MarriageAbroadCalculator` to encapsulate all the logic
required by the flow and views. Delegating to the
`MarriageAbroadDataQuery` for this method is a step toward that goal.

I'm sure the name of this method can be improved but I'm not 100% sure
what it means at the moment, so the renaming will have to wait.