Add #same_sex_ceremony_country_unknown_or_has_no_embassies?
I want the `MarriageAbroadCalculator` to encapsulate all the logic
required by the flow and views. Delegating to the
`MarriageAbroadDataQuery` for this method is a step toward that goal.

I'm sure the name of this method can be improved but I'm not 100% sure
what it means at the moment, so the renaming will have to wait.

For reference, the `#ss_unknown_no_embassies?` method was introduced in
6a9a7b84b1af57b9d6d263a18dc5611400daa4e4.