Update next_node block in check-uk-visa
In the PR adding Kuwait to the electronic visa waiver program, this 
piece of code was a little tricky to follow, so this commit is just
trying to improve that.

All conditional checks relating to the current question are moved to
the highest levels within in the next_node block, and any 
sub-conditional checks related to each purpose of visit are nested
within these.