WIP: Automatically identify possible next nodes
Previously when using `next_node` with a block, it was necessary to specify the
permitted next nodes. This was so that we could allow the flexibility of the
block to determine the next node, but still be able to visualise the possible
routes through the flow.

After this change, if you do not specify a set of permitted next nodes then you
have to specify the next nodes by calling the recently introduced `goto_node`
method inside the `next_node` block.

This is achieved by obtaining the source code of the `next_node` block, parsing
the source code into an AST, and then processing that AST to find calls to
the `goto_node` method.

Although I've added the `method_source` & `parser` gems to the `Gemfile`, they
were already implicit dependencies via the `pry` gem.

Before:

    multiple_choice :question? do
      option :option_one
      option :option_two

      permitted_next_nodes = [
        :outcome_one,
        :outcome_two
      ]

      next_node(permitted: permitted_next_nodes) do |response|
        case response
        when :option_one
          :outcome_one
        when :option_two
          :outcome_two
        end
      end
    end

After:

    multiple_choice :question? do
      option :option_one
      option :option_two

      next_node do |response|
        case response
        when :option_one
          goto_node :outcome_one
        when :option_two
          goto_node :outcome_two
        end
      end
    end

TODO:

* Do some work to see whether these changes have a detrimental impact on
performance. Since the parsing of the `next_node` blocks is done at flow
definition-time and since flow instances are currently cached in the
`FlowRegistry`, I think the only real impact should be at app start-up time.

* One possible performance improvement would be to re-use a single instance of
the parser and just reset it each time we need to use it.

* Improve variable naming it's a bit confusing since there are two types of
node: AST nodes and flow nodes.

* Add more unit tests for the AST processing and make the code more robust.

* Add an integration test for the new style of `next_node` block.

* I think it would be nicer if both the exceptions raised from within the
processing of a call to `next_node` were an `ArgumentError`.

* I think it would be nicer if the existing `ArgumentError` had an explanatory
message.