WIP: Introduce a goto_node method inside next_node blocks
Previously a `next_node` block had to return a node name symbol in order to
specify which node to go to next. In some places this means we've used
`next :node_name` in order to "jump" out of the block with a return value.

By introducing the `goto_node` method, I hope we can standardize on using this
method everywhere to make the code more readable. Also I think by using calls
to this new method consistently, we may be able to automatically detect the
possible next nodes, thereby avoiding the need to specify
`permitted_next_nodes` which is cluttering up the code quite a bit.

I'm using Ruby's `throw` & `catch` methods to implement this which isn't ideal,
but I think the advantages outweigh the disadvantages.

TODO:

* I've had to call `dup` on the `current_state` before extending it with the
`GotoNodeMethods` module, because it can sometimes be frozen. I need to
investigate why the state is only sometimes frozen and/or whether we could
freeze it after extending it with the module to avoid having to call `dup`.

* Add some kind of integration test using the new style `next_node` block.