Use instance_eval instead of eval for use_shared_logic
This change reduces the scope of the context in which the code in the shared
logic files is executed which feels like a good thing to do.

This change also supplies a file path and a line number to the call to
`instance_eval` which means we get much better error messages when there is a
`SyntaxError` in a shared logic file.