Use instance_eval instead of eval for shared logic
This means we no longer need to explicitly specify the `binding` which seems
a bit simpler.

Also, since the DSL "runs" against the instance of a flow, I think this code is
a bit more readable.

Note that I'm confident that this new implementation is exactly equivalent to
the previous implementation.