Don't allow multiple calls to next_node per question
I'm pretty sure we don't have any places where `next_node` is called multiple
times per question and I can't see any sensible justification for allowing it.

Constraining the DSL so that it can only be called once means that we can
simplify the code and make it easier to reason about.

With this guard condition in place, there's no longer any need to add the
`permitted_next_nodes` to any stored in a previous call.