Fix rubocop violation
```
lib/smart_answer/flow.rb:27:7: W: Lint/Eval: The use of eval is a serious security risk.
      eval File.read(path), binding, path.to_s
      ^^^^
```

This also means we no longer need to explicitly specify the `binding` which
seems a bit simpler.

Also, since the DSL "runs" against the instance of a flow, I think this code is
a bit more readable.

Note that I'm confident that this new implementation is exactly equivalent to
the previous implementation.