Ignore rubocop violation
```
lib/smart_answer/flow.rb:27:7: W: Lint/Eval: The use of eval is a serious security risk.
      eval File.read(path), binding, path.to_s
      ^^^^
```

This violation has only been triggered, because I've made a change to the line
which includes the call `eval` - the call to `eval` was there already i.e. I
haven't made anything worse!

I did experiment [1] with using `instance_eval` instead, but this caused
problems with constant lookup, so for the moment I think disabling the specific
Rubocop rule is the pragmatic solution.

[1]: #2321