Add calculator class to Overseas Passports
Extract a `book_appointment_online?` method from the partials and into a
Ruby class that can be used by the smart answer flow and passed to
partials.

At the moment, I’m leaving in the `current_location` checks within the
partials, even thought the value is now available on calculator class,
to a). to keep the commit small, and b). avoid those calls all together
and instead add richer methods that don’t require lists of countries to
be defined in partials.