Remove local current_location variable in partial
`current_location` can now be read from the overseas passports
calculator so this is the first step at removing this variable being passed around the smart answer files by hosing it out of partial `making_your_application`.

In future commits, we can further reduce calls to this variable with
methods that return true/false for conditions instead of having data
embedded in the partials to check against `current_location`.