Remove `save_input_as` state in overseas passports
To further remove the propagation of `current_location` throughout
overseas passports, this makes the `calculator` available in the first
question and sets `current_location` using a `next_node_calculation`
so `save_input_as ‘current_location’` can be removed.

The `next_node_calculation` setting the value of `current_location` can
be removed in future commits once the reliance on this variable is
removed.