Remove `data_query` instance from outcome template
Extend the overseas passports calculator to accept a data_query 
instance via the constructor, or create one if not received.

Add method `cash_only_country?` which delegates to the data_query
instance and allows us to remove the reference to the data_query in
outcome and partial and instead check this value on the calculator.