Remove `location` variable from flow
With `world_location` now available on the overseas passports
calculator class, this variable is no longer needed within the smart
answer flow or outcomes.

Checking `world_location` is still valid is moved to a validate block
which will keep the user on the same page if they enter an invalid
value.