Remove `application_type` variable from flow
Add `application_type` method in overseas passport calculator so it
can be tested with a unit test.

As the list of countries we allow the user to choose from in Q1 comes
from the GDS API, and `passport_data` is based on that countries
being present in data file `passport_data.yml`, a guard statement
is used to protect against these two sources falling out of sync.