Remove `ips_number` variable from flow
Add `ips_number` method to the smart answer calculator so we can put
a unit test around it and remove calculate block setting its value
in the smart answer flow.

Update outcomes and partials to retrieve values from the calculator.

For now, partials in the cost/ don’t receive the calculator, so am
leaving these as they are, receiving an ips_number rather than
calculator for now.