Remove `supporting_documents` variable from flow
Move the `supporting_document` calculation from the overseas passports
flow to the calculator class and add unit tests.

The value returned by `supporting_documents` changes if the user visits
the birth_location question and enters a value that is not the united
kingdom, so to cover the scenario where the method is invoked on a
calculator without this value set, we add a `blank?` check to return
a correct default value.

With the `ips_docs_number` value based on the return value of
`supporting_documents`, the calculation of `ips_docs_number` is
moved to the end of the flow where the user has entered their birth location if routed to that question.