Wrap question with article html tag
This commit adds the same article tag used in rendering outcomes
and landing pages to properly inherit CSS to allow for a more uniform
rendering in the questions. These tags wraps the former current question
tags and allow for proper govspeak rendering. This requirement raised to
cater for callouts within questions. This may expand in the future.