Introduce `permitted: :auto` option for next_node
For the moment, using this option sets `permitted_next_nodes` to be an empty
array. While this isn't immediately useful, it paves the way for automatically
detecting possible next nodes within a `next_node` block.

The new option is currently unusable, because if `permitted_next_nodes` is an
empty array then any result from the `next_node` block will result in an
exception since it won't be in the list of `permitted_next_nodes`.