Introduce parser for identifying method invocations
Although this is not currently used, I plan to use this to automatically
detect possible next nodes in a `next_node` block.

Although this is a fairly general solution, it does not currently handle
methods with negative arity i.e. it will just ignore invocations of them.

The `invocations` method returns an array of hashes each of which has two
keys. The `:method` key references the method object in the
`methods_of_interest` passed into the constructor. The `:arg_nodes` key
references an AST fragment representing the arguments with which the method
was called.

This commit introduces explicit dependencies on the `parser`, `ast` & `method_source` gems. However, `parser` was already an implicit dependency via `rubocop` and `ast` was already
an implicit dependency via `rubocop` and thence `parser`. `method_source` was already an
implicit dependency via `pry`. Thus no new actual dependencies have been introduced.

I've included `AST::Sexp` into the test in order to simplify the code when
asserting against a specific AST fragment.