Automatically identify permitted next nodes
This makes use of `SmartAnswer::Question::NextNode::Parser` to identify calls
to the syntactic sugar methods in
`SmartAnswer::Question::NextNode::InstanceMethods` i.e. `#question` and
`#outcome`.

Currently there's nothing forcing all return values to be wrapped in a call
to one of these methods, but I plan to address that in subsequent commits.