Move permitted next node auto-detection into reader
This is just a refactoring, i.e. there should be no change in behaviour.
Hence there are no changes to the tests.

Now that the automatic detection of permitted next nodes happens inside the
reader method, I've had to start calling the reader method from within
`Question::Base` in a few places.

My plan for future commits is to remove calls to the reader method from
within `Question::Base`. The only other places which call this method
are in `GraphPresenter`.

This will mean that the automatic detection of permitted next nodes only
happens when strictly necessary i.e. when visualising the flow and not
during normal user interaction.

The benefit of doing this is that we shouldn't need to worry about the
performance of the automatic detection code.