Relax permitted next node check when `permitted: :auto`
This change removes one of the places which was invoking the
`permitted_next_nodes` reader method from within `Question::Base`. We now
access the `@permitted_next_nodes` instance variable directly, avoiding
needlessly triggering the automatic detection code in the reader method.

Although it's nicer to fail-fast at flow-definition-time if no nodes are
returned via the syntactic sugar methods, an exception will still be raised
at flow-execution-time, because the chosen next node will not be in the
(empty) list of permitted next nodes.

I did contemplate moving the check into the `permitted_next_nodes` reader
method, but since I'm aiming for that to only be invoked from the
`GraphPresenter`, I'm not sure that would gain us much.

I also contemplated removing the check altogether (i.e. when auto-detection
is disabled). However, I think it's useful enough that the extra complexity
is worth it.