Alternative check for next node when `permitted: :auto`
Previously we always checked whether the node key returned from `next_node`
was in the list of permitted next nodes. I think that the main point of this
check was to ensure that the list of permitted next nodes stayed up-to-date
with the possible return values from `next_node` for each question, so that
the data used for visualisation is up-to-date.

The point of this commit is to avoid having to invoke the permitted next node
automatic detection code just in order to perform this check. I'm avoiding
that by calling the `@permitted_next_nodes` instance variable direct instead
of the reader method. Now the only place which uses the reader method is the
`GraphPresenter` used in visualisation.

The syntactic sugar methods, `question` & `outcome` now return the node key
wrapped in a new class, `NextNodeBlock::PermittedNodeKey` which allows us to
detect whether the node key returned from `next_node` has come from one of
the syntactic sugar methods. By raising an exception if it isn't we can be
sure that the automatic detection parsing will identify that node key as a
possible next node and hence ensure that the data used for visualisation is
kept up-to-date.

I've inherited `PermittedNodeKey` from `SimpleDelegator` so that to almost all
intents and purposes it "looks" just like a symbol. However, I noticed that
the `nil?` method is not delegated automatically. I've added unit tests to
assert the behaviour of methods which are being used in
`Question::Base#next_node_for` i.e. `present?`, `to_sym` & `to_s` (String
interpolation in exception message). Since Rails implements `Object#present?`
as the inverse of `Object#blank?`, I've tested the latter as well.

Although the code in `Question::Base` is now considerably more complicated than
I would like, much of that complexity is related to dealing with the "legacy"
style of `next_node` blocks, i.e. where permitted next nodes are explicitly
passed in. Since the idea is that this will go away, I'm confident that it'll
be possible to simplify the code significantly when we've converted all the
flows to use the new style of `next_node` blocks.