Improve unit test coverage for SmartAnswer::State
There have been some changes to the implementation of `OpenStruct` in
Ruby v2.3. These tests capture some of the behaviour that we are relying on.
They should make it easier to diagnose and fix the problems that the upgrade
will introduce. They all pass at this point i.e. under Ruby v2.2.4.