Fix SmartAnswer::State#method_missing for Ruby v2.3
Some changes have been made to the implementation of `OpenStruct` in Ruby v2.3:

* Define OpenStruct attributes lazily [1]
* Finish defining OpenStruct attributes lazily [2]

This means that attribute methods are now not defined until they are called for
the first time. This lazy definition of attribute methods is achieved in
`OpenStruct#method_missing`. This in turn means that `State#method_missing` is
now invoked when one of these attributes is accessed for the first time - it
wasn't being invoked previously. Our implementation of `State#method_missing`
was only calling `super` for writer methods and so incorrectly raised a
`NoMethodError` exception when one of the default attributes was accessed for
the first time.

I've changed the implementation of `State#method_missing` so that it uses
`OpenStruct#respond_to_missing` to check whether or not the attribute exists.
This seems to fix all the failing tests.

[1]: https://github.com/ruby/ruby/pull/1033
[2]: https://github.com/ruby/ruby/pull/1037