Avoid warning in WorldLocationTest with Ruby v2.3.0
After the upgrade to Ruby v2.3.0, the following warning started to appear in
the test output:

    $ ruby test/unit/world_location_test.rb -n "/return nil if not found/" -v
    Run options: -n "/return nil if not found/" -v --seed 37197

    # Running:

    [warning] The response contained in an RestClient::Exception is now a RestClient::Response instead of a Net::HTTPResponse, please update your code
    WorldLocationTest#test_: finding a location by slug should return nil if not found.  = 0.02 s = .

    Finished in 0.025900s, 38.6108 runs/s, 38.6108 assertions/s.

    1 runs, 1 assertions, 0 failures, 0 errors, 0 skips

The problem seems to occur when a 404 response is turned into an exception by
the `rest-client` gem and then that exception is rescued in the
`gds-api-adapters` gem. In particular it seems to be related to the lines
where the `http_body` on the exception is converted into JSON for logging [1].

I'm pretty sure this problem is related to this [issue][2] in `rest-client`,
but reading the comments, it sounds as if there's not going to be a fix until
v2.0.0 is released. We're currently using the latest in the 1.x series, v1.8.

As far as I can tell the problem isn't fixed in newer versions of
`gds-api-adapters` either.

So I've worked around the problem by stubbing at the `Services.worldwide_api`
level just for the relevant test. Personally I think this is a better level to
stub at rather than using the `gds-api-adapters` helper methods which use
`WebMock` to stub at a very low level. Ideally we'd change all the tests in
`WorldLocationTest` to work this way, but that would be a fair bit more work.

[1]: https://github.com/alphagov/gds-api-adapters/blob/v25.1.0/lib/gds_api/json_client.rb#L266-L267
[2]: https://github.com/rest-client/rest-client/issues/454