Fix SmartAnswer::State#respond_to? behaviour
It's good practice to implement `#respond_to_missing` whenever you implement
`#method_missing` on a class. I'm fixing this as part of the PR to upgrade Ruby
to v2.3, because we ran into some changes in `OpenStruct` which caused
unexpected exceptions in tests, and having this behaviour working correctly
will help us diagnose those problems.