Update consular fee in marriage abroad outcome
The FCO has clarified that any affirmation to marry should be £55 and 
variances such as this one ending in “in English” may need be replaced 
by the standard ‘Affirmation for marriage’.

This commit address the first issue, bringing the fee inline with all
other outcomes showing ‘Affirmation for freedom to marry in English’ 
and will allow us to now consolidate all fees in one partial.

Hopefully, having all the fees in one partial will allow us to address
the second issue, and allow further consolidation of consular fees and
eliminate unnecessary variances from appearing in marriage abroad
outcomes.