Add consular service fees table to marriage abroad
This commit introduces a partial for rendering consular services
and their corresponding fees, as shown to users in many of the marriage
abroad outcomes.

An outcome can use this partial by passing an array of services to
show the user in the call to render the partial, and the partial in
turn will render each service in the order it was received.

As the first commit of this PR, the fees from partial
`_consular_cni_os_fees_incl_null_osta_oath_consular_letter` are added
to a `consular_fees_table_items`, and is removed. All calls to render
the removed partial are replaced with calls to render the new fees
table.