Remove unnecessary regression test checksum file
This flow uses a newer style of testing and so doesn't have a regression test
associated with it.

A flow only has a regression test if it has a
`*-responses-and-expected-results.yml` file associated with it [1] and this
flow does not have such a file.

I believe this checksum file was accidentally added in this commit [2].

[1]: https://github.com/alphagov/smart-answers/blob/aabf580f3fbf53d6cdd4475380cb197d6dcacaea/lib/smart_answer_test_helper.rb#L10-L12
[2]: e757b611f3c1bae61ab26d9fdaf642ad0c0bb3c4