Add script to generate permitted next nodes for all flows
I'm about to change all the flows to use the  mechanism for automatically
detecting permitted next nodes. Although the existing tests, particularly
the regression tests should give a lot of confidence that nothing gets
broken, I'm hoping to compare the results of this script before and after
the changes as a double-check.