Remove incorrect permitted next node from check-uk-visa
The `outcome_visit_waiver` is not a valid return value for the `next_node`
block for this question.

I came across this while double-checking that the mechanism for automatic
detection of permitted next nodes was working correctly. The fact that this
had managed to get out of sync highlights the duplication inherent in the
current implementation and encourages me that the automatic detection
mechanism will be an improvement.