Replace precalculated variable in energy grants calc.
Add precalculated variable `under_green_deal` to the flow which is the
inverse of the boolean condition evaluated in
`qualified_for_help_from_energy_supplier` and remove the later from
the smart answer.

Having the inverse condition allows use to use an `if` condition
in the two places it's required to render under the green deal
content in the outcome, rather than one if and one unless were
we to use `qualified_for_help_from_energy_supplier`.