Split CalculateRedundancyPayTest into two tests
This integration test was testing two flows, albeit ones with a significant
amount of shared logic. It seems less surprising to have one test per flow
as we do elsewhere.

There are a bunch of indentation changes in this commit, so it's better to
view it using the following command:

    $ git show -p --ignore-space-change <sha>

I plan to fix the filename of the original test in a separate commit to keep
the diff as simple as possible.