Reduce duplication in receiving_non_exemption_benefits?
By changing the `calculate` block to a `next_node_calculation` block, we can
make the value available to the `next_node` block.

This in turn means we can avoid duplicating the logic to split the response and
construct the next question node key i.e. we can use
`benefit_related_questions.shift` in the `else` block.

This also has the benefit of making the `receiving_non_exemption_benefits?`
question more similar to the other "benefit amount" related questions.