Auto-detect permitted next nodes in benefit-cap-calculator
By using case statements to convert from benefit type to the relevant benefit
amount question, we can make the `next_node` logic less "dynamic" i.e. we can
explicitly enumerate all the possible results from the `next_node` blocks.

I'm not really very happy with this solution, but I don't think it's any worse
than the previous implementation and doing this means that we should be able to
remove the logic which deals with permitted next nodes that are explicitly
passed into a `next_node` block.

I'm not convinced that the visualisation for this flow is actually that useful,
particularly regarding the optional benefit amount questions. We could probably
get something that was almost as useful by extracting the "full" case statement
from the first of these questions into a method and re-using it in each
question. This would imply that you could always get to any other benefit amount
question from any benefit amount question. Obviously this isn't actually
accurate, but the visualisation might be good enough.

I think it would also be worth considering whether the user experience for this
flow is actually that good e.g. it might be better to group all the benefit
amount questions onto a single page.