Add #civil_partnership_institution_name
I'm refactoring marriage-abroad so that the `MarriageAbroadCalculator`
contains everything necessary to render the outcomes. Adding this method
will allow me to remove the `:institution_name` in a subsequent commit.

I've added the `civil_partnership` prefix to the new method name as the
value is only displayed in a civil partnership outcome.