Inline :three_day_residency_handled_by_exception
I _think_ this makes it clearer that the countries in this array
are related to the countries listed in the conditionals just before this
block of code in outcome_os_consular_cni. It also makes it clearer that
'italy' shouldn't be in this list as that now has its own outcome.