Inline :three_day_residency_handled_by_exception
Note. We can ignore Italy in this country list as opposite sex
ceremonies in Italy are handled by the outcome_os_italy outcome..

There are a few places in the outcome_os_consular_cni template that deal
with ceremonies in Croatia and Russia. Inlining this condition makes it
clearer that this is related to those other conditions in this template.
I think this'll make it easier to remove this duplication in future.