Remove unreachable code from SmartAnswer::Question::Base
The only way for `@permitted_next_nodes` to *not* be `:auto` is when `next_node`
is called without a block i.e. when a single node key is specified.

In this scenario there is no way that the internally-defined block can return
with anything other than the single node key in the `@permitted_next_nodes`.

Thus the exception will never be raised.