Add #birth_certificate_required_as_supporting_document?
This is confusing as the NO_BIRTH_CERT_REQUIREMENT array contains a list
of countries that don't require a birth certificate as a supporting
document. We then check that the ceremony country isn't in that list in
order to determine whether a birth certificate is required. I'm sure this
can be improved but I think it's OK for now.