Move Macedonia condition out of :notary_public_inclusion
I _think_ Macedonia is in this list because it's handled elsewhere in
the outcome_os_consular_cni template, rather than because it's a
`:notary_public_inclusion` country.

There are already other special cases dealing with Macedonia in the
template and I'm hoping this'll make it easy to remove more duplication
further down the line.