Add #document_download_link_if_opposite_sex_resident_of_uk_countries?
Use it instead of the
`:no_document_download_link_if_os_resident_of_uk_countries` precalculate
block.

Part of refactoring marriage-abroad. See doc/refactoring.md for more.

I struggled with the naming of this method as the logic seems confusing.
The `NO_DOCUMENT_DOWNLOAD_LINK_IF_OS_RESIDENT_OF_UK_COUNTRIES` array
contains a list of ceremony countries that _don't_ allow you to download
the notice of and affidavit for marriage forms. We check that the
ceremony country _isn't_ in this list in order to decide whether to
display the download