Rename outcome_os_italy
I used the marriage-abroad-responses-and-expected-results.yml file to
determine the responses that lead to this outcome.

This outcome is displayed for opposite sex marriages in Italy when the
applicant resides in either the UK or Italy.

Opposite sex marriages in Italy when the applicant resides in a third
country are handled by the
outcome_consular_cni_os_residing_in_third_country outcome.