Rename outcome_os_no_cni
I used a combination of the
marriage-abroad-responses-and-expected-results.yml file and outcome
template to confirm that this outcome is used for for opposite sex
marriages in no-CNI countries when the applicant resides in either the
ceremony or a third country.

NOTE. This is true for all countries except st-martin and
saint-barthelemy, where this outcome is also displayed for UK residents,
and for same sex ceremonies. I'm going to investigate that separately as
it doesn't seem quite right.