Rename outcome_os_no_cni
I used the marriage-abroad-responses-and-expected-results.yml file to
determine the responses that lead to this outcome.

This outcome is displayed for opposite sex marriages including residents
of the ceremony country or a third country in "no CNI" countries.

NOTE. This is true for all countries except st-martin and
saint-barthelemy, where this outcome is also displayed for UK residents,
and for same sex ceremonies. I'm going to investigate that separately as
it doesn't seem quite right.