Rename outcome_os_other_countries
I used the marriage-abroad-responses-and-expected-results.yml file to
determine the responses that lead to this outcome.

This outcome is displayed for all opposite sex marriages in a number of
countries.

It looks as though it might not be too difficult to split this outcome
into country specific outcomes, although that'll have to wait for
another time.