Rename outcome_os_other_countries
I used a combination of the
marriage-abroad-responses-and-expected-results.yml file and outcome
template to confirm that this outcome is used for for opposite sex
marriages in various countries.

It looks as though it might not be too difficult to split this outcome
into country specific outcomes, although that'll have to wait for
another time.