Rename outcome_cp_france_pacs
I used the marriage-abroad-responses-and-expected-results.yml file to
determine the responses that lead to this outcome.

This outcome is displayed for civil partnerships in France or French
overseas territories.

I've added 'or_french_overseas_territory' to mirror the naming of
outcome_opposite_sex_france_or_french_overseas_territory.

I've removed 'pacs' from the name as I believe that's equivalent to
'civil_partnership', and we're already using 'civil partnership'
elsewhere.