Rename outcome_cp_france_pacs
I used a combination of the
marriage-abroad-responses-and-expected-results.yml file and outcome
template to confirm that this outcome is used for for civil partnerships
in France and French overseas territories.

I've added 'or_french_overseas_territory' to mirror the naming of
outcome_opposite_sex_in_france_or_french_overseas_territory.

I've removed 'pacs' from the name as I believe that's equivalent to
'civil_partnership', and we're already using 'civil partnership'
elsewhere.