Rename outcome_cp_all_other_countries
I used the marriage-abroad-responses-and-expected-results.yml file to
determine the responses that lead to this outcome.

This outcome is displayed for same sex marriages and civil partnerships
in "other countries".

I've changed 'all_other_countries' to 'other_countries' to mirror the
naming of 'outcome_opposite_sex_other_countries'.

NOTE: This is essentially a duplicate of
outcome_same_sex_marriage_not_possible but I'm going to investigate that
separately.