Rename outcome_cp_all_other_countries
I used a combination of the
marriage-abroad-responses-and-expected-results.yml file and outcome
template to confirm that this outcome is used for for same sex marriages
and civil partnerships in various countries.

I've changed 'all_other_countries' to 'other_countries' to mirror the
naming of 'outcome_opposite_sex_other_countries'.

NOTE: This is essentially a duplicate of
outcome_same_sex_marriage_not_possible but I'm going to investigate that
separately.