Rename outcome_ss_marriage_not_possible
I used a combination of the
marriage-abroad-responses-and-expected-results.yml file and outcome
template to confirm that this outcome is used for for same sex marriages
and civil partnerships when they're not possible.

NOTE: This is essentially a duplicate of
outcome_same_sex_marriage_and_civil_partnership_other_countries but I'm
going to investigate that separately.