Exclude existing Rubocop violations using auto-generated config
Previously we were using the `--diff --cached` options with `govuk-lint-ruby`
to only run the Rubocop checks against code that had been changed in the given
branch i.e. it's a way to avoid having to fix all the violations up-front.

However, I'm not convinced this is a very sensible approach, because it means
we fix the violations in rather a piecemeal way. Also I think we've run into a
catch-22 situation a couple of times where we wanted to revert a merge commit
to put the code back *exactly* as it was previously. The problem is that the
revert commit can draw Rubocop's attention to "old" lines of code that may not
pass the Rubocop checks - thus preventing us from doing the straightforward
revert.

In this commit I've generated a project-specific `.rubocop_todo.yml` file using
`govuk-lint-ruby --auto-gen-config --exclude-limit 99999` which excludes
existing violations on a per-cop basis [1]. The limit was set to 99999 to
minimise the number of cops that are disabled entirely - this way most cops are
only disabled for a specific set of files.

The generated `.rubocop_todo.yml` file is included into the configuration via a
local `.rubocop.yml` file using an `inherit_from` key. This means that the
exclusions in `.rubocop_todo.yml` are used for all `govuk-lint-ruby` commands
by default. Note that the `.rubocop.yml` file does not *override* the cop
definitions in the `govuk-lint` gem - it is merely a way to add *extra*
configuration.

Using this approach should prevent us from introducing new violations into the
project without needing to fix all the existing violations first. However, in
the long run, the idea is for us to gradually remove these exclusions to bring
the project fully in line with the default `govuk-lint-ruby` configuration.

This should be pretty straightforward, since we can just fix violations of a
single cop or even just violations of a single cop in a single file and then
update the `.rubocop_todo.yml` file accordingly.

Ultimately we will be able to remove the `.rubocop_todo.yml` & `.rubocop.yml`
files and just rely on the default `govuk-lint` configuration.

Note that as per the header comment in `.rubocop_todo.yml`, changes in the
inspected code, or installation of new versions of RuboCop, may require
`.rubocop_todo.yml` to be generated again.

I've also updated the `jenkins.sh` CI script and the relevant documentation.

[1]: https://github.com/bbatsov/rubocop/blob/master/README.md#automatically-generated-configuration