Avoid double render error if exception raised after render
The `if performed?` conditional in `ApplicationController#error` was introduced
in this commit [1] in order to avoid a `DoubleRenderError` if the relevant
exception was raised after rendering had completed. However, the change was
made along with a test that didn't test what it was meant to test - I think the
test is triggering an exception *in the middle of* rendering and not *after*
rendering.

To remedy this, I replaced the originally added controller test and added a new
Rails integration test, `ErrorHandlingTest`. This gives coverage of both
branches of the `if/else` statement and a failure of one of these tests
surfaced the fact that the line, `response_body = error_message` is not doing
anything useful - it's just assigning the value in `error_message` to a local
variable, `response_body` which is not then used for anything.

In fact I first came across this because this method was causing a
`Lint/UselessAssignment` Rubocop violation where a `response_body` local
variable was being assigned, but not used. I _think_ the intention was probably
to use the `ActionController::Metal#response_body` writer method.

I fixed the failing test by changing `response_body = error_message` to
`self.response_body = error_message`.

I also decided to change `response.status = status_code` so that it too uses a
writer method on `ActionController::Metal` rather than accessing the
`response`. The latter is only made available by
`ActionController::RackDelegation` i.e. it may not be available in all
requests. Doing this felt more consistent and more robust.

I had to use `set_slimmer_headers(skip: true)` in the test controller action,
because Slimmer ignores the body content of an error page and so we'd never see
what it's being set to. The body content is ignored because
`Slimmer::Skin#error` does not include `Slimmer::Processors::BodyInserter` in
the list of processors for the request.

Thus with Slimmer enabled the relevant ERB error template (e.g. 404.html.erb
[2] and its related partial [3]) from the static app is the main source of the
error page, albeit with a few things changed e.g. the page title.

[1]: 168b69edf25c2a816be2b947c93c4d6630b3f6a6
[2]: https://github.com/alphagov/static/blob/master/app/views/root/404.html.erb
[3]: https://github.com/alphagov/static/blob/master/app/views/root/_error_page.html.erb