Wrap case statement in begin/end to avoid Rubocop violations
The Style/CaseIndentation Rubocop rule requires that `when` statements are
indented at the same level as their parent `case` statement. However, in this
case we are assigning the result of the `case` statement to a local variable
which pushes the `case` keyword to a deep level of "indentation".

I think rigidly applying the rule in this case would lead to code that was hard
to read, so I've elected to wrap this `case` statement in a `begin` ... `end`
block. This means that the `case` statement can be at a sensible level of
indentation.

Previously everything inside the `case` statement was indented an odd number of
characters which was odd. This commit also fixes that problem.