Exclude tmp directory from Rubocop checks
By default Rubocop looks for Ruby files in all project directories. There seem
to be some Ruby files relating to `govuk-content-schemas` in the project `tmp`
directory on the Jenkins CI server which contain cop violations. We don't care
about files in this directory, so we can exclude them from the Rubocop checks.