Remove unit tests for use_legacy_data option
This `country_select` option is not used in any of the "production" flows and so
there's no point in testing it.

Note that I had to stub the Worldwide API in `FlowTest` now that the test is
using a standard `country_select`, but this is a good thing, because the test
is now more realistic.