Update outcome_same_sex_marriage_and_civil_partnership_not_possible
Replace "your same-sex" with "a same-sex". This makes this outcome
identical to
outcome_same_sex_marriage_and_civil_partnership_in_other_countries which
will make it easier to remove the duplication.

I chose to update this outcome to "a same-sex" (rather than updating the
other outcome to "your same-sex") because we're already using "a
same-sex" in other templates, and "your same-sex" was only used in this
outcome.

This outcome is shown when
`MarriageAbroadDataQuery#ss_marriage_not_possible?` returns `true`.
Expanding the logic of that method indicates that this change will
affect:

* Ceremonies in Hungary, Mongolia, Montenegro, Nicaragua and Russia when
  your partner is not british

* Ceremonies in Azerbaijan, Chile, Dominican Republic, Kosovo, Latvia,
  Cambodia and Vietnam when your partner is a national of the ceremony
  country.

* Ceremonies in San Marino and Seychelles

I've run the regression tests to update the artefacts after making this
change.